requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==21.4.0
bcrypt==3.2.2
billiard==3.6.4.0
black==22.6.0
cached-property==1.5.2
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
cryptography==37.0.4
Deprecated==1.2.13
distro==1.7.0
Django==4.0.6
django-celery-beat==2.3.0
django-decouple==2.1
django-elasticsearch-dsl==7.2.2
django-js-asset==2.0.0
django-mptt==0.13.4
django-ninja==0.19.1
django-redis==5.2.0
django-timezone-field==5.0
djangorestframework==3.13.1
docker==4.2.2
docker-compose==1.26.2
dockerpty==0.4.1
docopt==0.6.2
elasticsearch==7.17.6
elasticsearch-dsl==7.4.0
factory-boy==3.2.1
Faker==13.15.0
flower==1.2.0
humanize==4.4.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
jsonschema==3.2.0
kombu==5.2.4
mypy-extensions==0.4.3
packaging==21.3
paramiko==2.11.0
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pydantic==1.10.2
PyNaCl==1.5.0
pyparsing==3.0.9
pypiwin32==223
pyrsistent==0.18.1
pytest==6.2.5
pytest-base-url==1.4.2
pytest-django==4.5.2
pytest-factoryboy==2.5.0
pytest-html==3.1.1
pytest-metadata==2.0.1
pytest-selenium==3.0.0
pytest-variables==1.9.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
pywin32==304
PyYAML==5.4.1
redis==4.3.4
requests==2.28.1
selenium==3.141.0
six==1.16.0
sqlparse==0.4.2
tenacity==6.3.1
texttable==1.6.4
toml==0.10.2
tomli==2.0.1
tornado==6.2
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.59.0
wrapt==1.14.1
